create separate LeaderboardClient
